# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a Social Network - OnlyFans api specifications API REST .",
    "version": "1.0.2",
    "title": "Swagger Api-profile",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "yannick.simon@kysoe.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/gallery": {
      "post": {
        "tags": [
          "gallery"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "gallery post",
        "description": "gallery post desccription",
        "operationId": "galleryPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Gallery"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Gallery"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gallery"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "gallery list",
        "description": "gallery desccription",
        "operationId": "galleryList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gallery"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gallery"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gallery/{gallery_id}": {
      "get": {
        "tags": [
          "gallery"
        ],
        "summary": "gallery get",
        "description": "gallery get desccription",
        "operationId": "galleryGet",
        "parameters": [
          {
            "name": "gallery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gallery"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Gallery"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gallery"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "gallery delete",
        "description": "gallery delete desccription",
        "operationId": "galleryDelete",
        "parameters": [
          {
            "name": "gallery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/picture": {
      "post": {
        "tags": [
          "picture"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "picture post",
        "description": "picture post desccription",
        "operationId": "picturePost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Picture"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Picture"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "picture"
        ],
        "summary": "picture list",
        "description": "picture desccription",
        "operationId": "pictureList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Picture"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Picture"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/picture/{picture_id}": {
      "get": {
        "tags": [
          "picture"
        ],
        "summary": "picture get",
        "description": "picture get desccription",
        "operationId": "pictureGet",
        "parameters": [
          {
            "name": "picture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Picture"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Picture"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "picture"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "picture delete",
        "description": "picture delete desccription",
        "operationId": "pictureDelete",
        "parameters": [
          {
            "name": "picture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/video": {
      "post": {
        "tags": [
          "video"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "video post",
        "description": "video post desccription",
        "operationId": "videoPost",
        "requestBody": {
          "$ref": "#/components/requestBodies/Video"
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "video"
        ],
        "summary": "video list",
        "description": "video desccription",
        "operationId": "videoList",
        "parameters": [
          {
            "in": "query",
            "name": "append_to_response",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "add additional data to response"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/video/{video_id}": {
      "get": {
        "tags": [
          "video"
        ],
        "summary": "video get",
        "description": "video get desccription",
        "operationId": "videoGet",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "video"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "video delete",
        "description": "video delete desccription",
        "operationId": "videoDelete",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "requestBodies": {
...
      "Gallery": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Gallery"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Gallery"
            }
          }
        },
        "description": "Gallery object that needs to be add",
        "required": true
      },
      "Picture": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Picture"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Picture"
            }
          }
        },
        "description": "Picture object that needs to be add",
        "required": true
      },
      "Video": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Video"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/Video"
            }
          }
        },
        "description": "Video object that needs to be add",
        "required": true
      },
...
    },
    "schemas": {
...
      "Gallery": {
        "type": "object",
        "required": [
          "profile_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "profile_id": {
            "type": "string"
          }
        }
      },
      "Picture": {
        "type": "object",
        "required": [
          "gallery_id",
          "filename"
        ],
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "filename": {
            "type": "string"
          },
          "gallery_id": {
            "type": "string"
          }
        }
      },
      "Video": {
        "type": "object",
        "required": [
          "gallery_id",
          "filename"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "gallery_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}